+++ .gitignore
@@ -59,3 +59,4 @@ site
 
 # node modules
 node_modules
+gen-ai/inference/nvidia-triton-server-gpu/yolo/triton-client/__pycache__
